pyproject.toml
[project]
name = "coding_papers_llm_assistant"
version = "0.1.0"
description = "LLM assistant for coding and papers"
requires-python = ">=3.12,<3.14"
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.1.0",
    "openinference-instrumentation-langchain>=0.1.56,<0.2.0",
    "opentelemetry-instrumentation-httpx>=0.1.0",
    "arize-phoenix>=12.15.0,<13.0.0",
    "pydantic>=2.0.0",
    "arxiv>=2.1.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
multi-agent-lab = "main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.poetry]
package-mode = false
